Move all color console display to nunit.common for reuse
