+++ src/NUnitCommon/nunit.common/TextDisplay/ColorStyle.cs
@@ -1,6 +1,6 @@
 ﻿// Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
-namespace NUnit.ConsoleRunner
+namespace NUnit.TextDisplay
 {
     /// <summary>
     /// ColorStyle enumerates the various styles used in the console display
