+++ src/NUnitConsole/nunit4-console.tests/ColorStyleTests.cs
@@ -3,7 +3,7 @@
 using System;
 using NUnit.Framework;
 
-namespace NUnit.ConsoleRunner
+namespace NUnit.TextDisplay
 {
     [TestFixture]
     public class ColorStyleTests
