+++ src/NUnitConsole/nunit4-console.tests/TestEventHandlerTests.cs
@@ -4,6 +4,7 @@ using System;
 using System.IO;
 using System.Text;
 using NUnit.Framework;
+using NUnit.TextDisplay;
 
 namespace NUnit.ConsoleRunner
 {
