+++ src/NUnitConsole/nunit4-console/ConsoleTestResult.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Xml;
+using NUnit.TextDisplay;
 
 namespace NUnit.ConsoleRunner
 {
